+++ src/jvmMain/kotlin/ehn/techiop/hcert/kotlin/chain/impl/SchemaValidationAdapter.kt
